+++ src/index.js
@@ -1,31 +1,11 @@
-'use strict';
+import {readFile, writeFile} from 'fs/promises';
 
-// Todo: Adapt this example code into a binary as part of its own library
-/**
- * @example
- * 'use strict';
- * const workboxBuild = require('workbox-build');
- * const findESResources = require('find-es-resources');
- *
- * (async () => {
- * const additionalManifestEntries = await findESResources('...file...');
- *
- * // Also use https://developers.google.com/web/tools/workbox/reference-docs/latest/module-workbox-routing.NavigationRoute
- * //  for the sake of Single-Page-Application `replaceState`/`pushState` use.
- * // See also https://medium.com/@george.norberg/history-api-getting-started-36bfc82ddefc
- * return workboxBuild.generateSW({
- *   additionalManifestEntries,
- *   swDest: 'sw.js'
- * });
- * })();
- */
+import path from 'path';
 
-const path = require('path');
+import {traverse as esFileTraverse} from 'es-file-traverse';
+import esquery from 'esquery';
 
-const {traverse: esFileTraverse} = require('es-file-traverse');
-const esquery = require('esquery');
-
-const queries = require('./queries.js');
+import queries from './queries.js';
 
 /**
 * @typedef {string} queryASTString
@@ -56,34 +36,64 @@ const parsedQueries = parseQueries(queries);
  */
 
 /**
- * @param {string} file
- * @param {external:EsFileTraverseOptions} esFileTraverseOptions
- * @param {{queryModule: string}} queryOptions
+ * @param {PlainObject} cfg
+ * @param {string} cfg.input
+ * @param {string} [cfg.removeBasePath=""]
+ * @param {string} [cfg.addBasePath=""]
+ * @param {external:EsFileTraverseOptions} [cfg.esFileTraverseOptions]
+ * @param {{queryModule: string}} [cfg.queryOptions={}]
  * @returns {Promise<string[]>}
+ * @example
+ * import workboxBuild from 'workbox-build';
+ * import findESResources from 'find-es-resources';
+ *
+ * (async () => {
+ * const additionalManifestEntries = await findESResources();
+ *
+ * // Also use https://developers.google.com/web/tools/workbox/reference-docs/latest/module-workbox-routing.NavigationRoute
+ * //  for the sake of Single-Page-Application `replaceState`/`pushState` use.
+ * // See also https://medium.com/@george.norberg/history-api-getting-started-36bfc82ddefc
+ * return workboxBuild.generateSW({
+ *   additionalManifestEntries,
+ *   swDest: 'sw.js'
+ * });
+ * })();
  */
-const findESResources = async (
-  file, esFileTraverseOptions, queryOptions = {}
-) => {
+const findESResources = async ({
+  input, removeBasePath, addBasePath,
+  esFileTraverseOptions = {}, queryOptions = {}
+} = {}) => {
   const esResources = new Set();
+  let queryModule;
+  if (queryOptions.queryModule) {
+    // eslint-disable-next-line no-unsanitized/method -- Runtime-specified
+    queryModule = (await import(
+      path.join(
+        process.cwd(),
+        queryOptions.queryModule
+      )
+    )).default;
+  }
+
+  if (!input) {
+    const pkg = JSON.parse(
+      await readFile(path.join(process.cwd(), './package.json'))
+    );
+    input = pkg?.exports?.browser;
+  }
+
   const filesArr = await esFileTraverse({
     node: true,
     ...esFileTraverseOptions,
-    file,
+    file: input,
     // excludePathExpression: '',
     callback (type, {ast}) {
       if (type !== 'enter') {
         return;
       }
 
-      const _parsedQueries = queryOptions.queryModule
-        // eslint-disable-next-line max-len -- Long
-        // eslint-disable-next-line import/no-dynamic-require, node/global-require -- Runtime-specified
-        ? [...parsedQueries, ...parseQueries(require(
-          path.join(
-            process.cwd(),
-            queryOptions.queryModule
-          )
-        ))]
+      const _parsedQueries = queryModule
+        ? [...parsedQueries, ...parseQueries(queryModule)]
         : parsedQueries;
 
       _parsedQueries.forEach(({parsedQuery, getPaths}) => {
@@ -106,7 +116,41 @@ const findESResources = async (
     esResources.add(_file);
   });
 
-  return [...esResources];
+  const ret = [...esResources];
+
+  if (removeBasePath || addBasePath) {
+    return ret.map((item) => {
+      return addBasePath + item.replace(removeBasePath, '');
+    });
+  }
+
+  return ret;
 };
 
-module.exports = findESResources;
+/**
+ * @param {PlainObject} cfg
+ * @param {string} cfg.output
+ * @param {string} [cfg.input]
+ * @param {string} [cfg.removeBasePath=""]
+ * @param {string} [cfg.addBasePath=""]
+ * @param {external:EsFileTraverseOptions} [cfg.esFileTraverseOptions]
+ * @param {{queryModule: string}} [cfg.queryOptions]
+ * @returns {Promise<string[]>}
+ */
+const saveESResources = async ({
+  output, input, removeBasePath, addBasePath,
+  esFileTraverseOptions, queryOptions
+}) => {
+  const resources = await findESResources({
+    input, esFileTraverseOptions, queryOptions
+  });
+  await writeFile(
+    output,
+    output.endsWith('.js')
+      ? 'export default ' + JSON.stringify(resources, null, 2) + ';\n'
+      : JSON.stringify(resources, null, 2) + '\n'
+  );
+  return resources;
+};
+
+export {findESResources, saveESResources};