+++ test/test.js
@@ -1,8 +1,10 @@
-'use strict';
+import {join, dirname} from 'path';
 
-const {join} = require('path');
+import {fileURLToPath} from 'url';
 
-const findESResources = require('../src/index.js');
+import {findESResources} from '../src/index.js';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
  * @param {string} file
@@ -14,7 +16,9 @@ function getFixturePath (file) {
 
 describe('findESResources', function () {
   it('Finds resources (no imports)', async function () {
-    const resources = await findESResources(getFixturePath('fetches.js'));
+    const resources = await findESResources({
+      input: getFixturePath('fetches.js')
+    });
     expect(resources).to.have.lengthOf(3);
     expect(resources).to.include.members([
       './test1.json',
@@ -23,8 +27,24 @@ describe('findESResources', function () {
     expect(resources[2]).to.match(/test\/fixtures\/fetches\.js/u);
   });
 
+  it('Finds resources (base paths)', async function () {
+    const resources = await findESResources({
+      input: getFixturePath('fetches.js'),
+      addBasePath: '/basePath/',
+      removeBasePath: './'
+    });
+    expect(resources).to.have.lengthOf(3);
+    expect(resources).to.include.members([
+      '/basePath/test1.json',
+      '/basePath/test2.json'
+    ]);
+    expect(resources[2]).to.match(/test\/fixtures\/fetches\.js/u);
+  });
+
   it('Finds resources (single import)', async function () {
-    const resources = await findESResources(getFixturePath('importer.js'));
+    const resources = await findESResources({
+      input: getFixturePath('importer.js')
+    });
     expect(resources).to.have.lengthOf(4);
     expect(resources).to.include.members([
       './test1.json',
@@ -35,11 +55,12 @@ describe('findESResources', function () {
   });
 
   it('Finds resources (queryModule)', async function () {
-    const resources = await findESResources(
-      getFixturePath('file-with-custom-items.js'), null, {
+    const resources = await findESResources({
+      input: getFixturePath('file-with-custom-items.js'),
+      queryOptions: {
         queryModule: './test/fixtures/queryModule.js'
       }
-    );
+    });
     expect(resources).to.have.lengthOf(4);
     expect(resources).to.include.members([
       './test1.json',